test: split test_app_registry.py by test class
The file was 999 lines, over the file-sizes check's 800-line threshold
(HSL102). Split by test class into two files, moving the two shared
SimpleNamespace factories into a family-scoped _fixtures_app_registry.py
re-exported from conftest.py, matching this directory's fixture convention.

Test bodies are unchanged.
